+++ tests/core/contracts/conftest.py
@@ -7,7 +7,7 @@ from eth_utils import (
 )
 import pytest_asyncio
 
-from _utils import (
+from tests.core.contracts.utils import (
     async_deploy,
     deploy,
 )
