+++ tests/core/contracts/test_contract_caller_interface.py
@@ -2,7 +2,7 @@ import pytest
 
 import pytest_asyncio
 
-from _utils import (
+from tests.core.contracts.utils import (
     async_deploy,
     deploy,
 )
