+++ tests/core/contracts/utils.py
