+++ tests/core/filtering/test_basic_filter_tests.py
@@ -1,6 +1,6 @@
 import pytest
 
-from tests.core.filtering._utils import (
+from tests.core.filtering.utils import (
     async_range,
 )
