+++ tests/core/filtering/test_filter_against_latest_blocks.py
@@ -1,6 +1,6 @@
 import pytest
 
-from tests.core.filtering._utils import (
+from tests.core.filtering.utils import (
     async_range,
 )
 from web3._utils.threads import (
